lab13k

+++ src/main/java/com/example/songr/modules/Song.java
@@ -1,12 +1,15 @@
 package com.example.songr.modules;
 
 import javax.persistence.*;
+import java.util.Optional;
 
 @Entity
 public class Song {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
+
 
     public Long getId() {
         return id;
@@ -15,16 +18,17 @@ public class Song {
     public void setId(Long id) {
         this.id = id;
     }
+
     @ManyToOne
     Album album;
 
     String title;
     long length;
-    short trackNumber;
+    int trackNumber;
 
     public Song(){}
 
-    public Song(String title, long length, short trackNumber, Album album){
+    public Song(String title, long length, int trackNumber, Album album){
         this.title = title;
         this.length = length;
         this.trackNumber = trackNumber;
@@ -55,11 +59,11 @@ public class Song {
         this.length = length;
     }
 
-    public short getTrackNumber() {
+    public int getTrackNumber() {
         return trackNumber;
     }
 
-    public void setTrackNumber(short trackNumber) {
+    public void setTrackNumber(int trackNumber) {
         this.trackNumber = trackNumber;
     }
 }
